+++ lib/superconfig.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module SuperConfig
-  VERSION = "2.2.1"
+  VERSION = "3.0.0"
 
   MissingEnvironmentVariable = Class.new(StandardError)
   MissingCallable = Class.new(StandardError)